T-Blade3
T-blade3 is a general parametric 3D blade geometry builder. The tool can
create a variety of 3D blade geometries based on few basic parameters
and limited interaction with a CAD system. The geometric and aerodynamic
parameters are used to create 2D airfoils and these airfoils are stacked
on the desired stacking axis. The tool generates a specified number of
2D blade sections in a 3D Cartesian coordinate system. The geometry
modeler can also be used for generating 3D blades with special features
like bent tip, split tip and other concepts, which can be explored with
minimum changes to the blade geometry. The use of control points for the
definition of splines makes it easy to modify the blade shapes quickly
and smoothly to obtain the desired blade model. The second derivative of
the mean-line (related to the curvature) is controlled using B-splines
to create the airfoils. This is analytically integrated twice to obtain
the mean-line. A smooth thickness distribution is then added to the
airfoil with two options either the Wennerstrom distribution or a
quartic B-spline thickness distribution. B-splines have also been
implemented to achieve customized airfoil leading and trailing edges.